+++ src/Kronika/Date/Trait/DateUnit.php
@@ -15,6 +15,7 @@ namespace Kronika\Date\Trait;
 
 use Kronika\Date\DateUnit as Unit;
 use Kronika\LocalDateTime;
+use Kronika\OverflowMode;
 use Kronika\Utils\RescueTrait;
 
 /**
@@ -31,8 +32,8 @@ trait DateUnit
 
     /** @internal {@see \Kronika\DateTime::with()} */
     #[\Override]
-    final public function _withinDateTime(LocalDateTime $datetime, bool $rolling): LocalDateTime
+    final public function _withinDateTime(LocalDateTime $datetime, OverflowMode $mode): LocalDateTime
     {
-        return $datetime->with($datetime->date()->with($this, $rolling), $rolling);
+        return $datetime->with($datetime->date()->with($this, $mode), $mode);
     }
 }