+++ src/Kronika/Time.php
@@ -647,7 +647,7 @@ final readonly class Time implements Unit
 
     /** @internal {@see DateTime::with()} */
     #[\Override]
-    public function _withinDateTime(LocalDateTime $datetime, bool $rolling): LocalDateTime
+    public function _withinDateTime(LocalDateTime $datetime, OverflowMode $mode): LocalDateTime
     {
         return $this->at($datetime->date());
     }